# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Site Studio multi-client website platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type SiteTemplate = typeof SiteTemplate[keyof typeof SiteTemplate];


export const SiteTemplate = {
  modern: 'modern',
  classic: 'classic',
  bold: 'bold',
  minimal: 'minimal',
  custom: 'custom',
} as const;

export type SiteSettingsSocialLinks = {[key: string]: string};

export interface SiteSettings {
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  faviconUrl?: string | null;
  /** @pattern ^#[0-9A-Fa-f]{6}$ */
  primaryColor: string;
  /** @pattern ^#[0-9A-Fa-f]{6}$ */
  accentColor: string;
  fontFamily: string;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /**
     * @nullable
     * @pattern ^G-[A-Z0-9]+$
     */
  ga4MeasurementId?: string | null;
  socialLinks?: SiteSettingsSocialLinks;
}

export interface Site {
  id: string;
  name: string;
  slug: string;
  template: SiteTemplate;
  settings: SiteSettings;
  updatedAt?: string;
}

export type SiteInputTemplate = typeof SiteInputTemplate[keyof typeof SiteInputTemplate];


export const SiteInputTemplate = {
  modern: 'modern',
  classic: 'classic',
  bold: 'bold',
  minimal: 'minimal',
  custom: 'custom',
} as const;

export interface SiteInput {
  /** @minLength 1 */
  name: string;
  /**
     * @minLength 1
     * @pattern ^[a-z0-9]+(?:-[a-z0-9]+)*$
     */
  slug: string;
  template?: SiteInputTemplate;
  settings?: SiteSettings;
}

export type SiteUpdateTemplate = typeof SiteUpdateTemplate[keyof typeof SiteUpdateTemplate];


export const SiteUpdateTemplate = {
  modern: 'modern',
  classic: 'classic',
  bold: 'bold',
  minimal: 'minimal',
  custom: 'custom',
} as const;

export interface SiteUpdate {
  /** @minLength 1 */
  name?: string;
  template?: SiteUpdateTemplate;
  settings?: SiteSettings;
}

export type BlockType = typeof BlockType[keyof typeof BlockType];


export const BlockType = {
  hero: 'hero',
  'text-image': 'text-image',
  features: 'features',
  testimonials: 'testimonials',
  faq: 'faq',
  cta: 'cta',
  gallery: 'gallery',
  contact: 'contact',
  'rich-text': 'rich-text',
  spacer: 'spacer',
} as const;

export type BlockData = { [key: string]: unknown };

export interface Block {
  id: string;
  type: BlockType;
  data: BlockData;
}

export type PageStatus = typeof PageStatus[keyof typeof PageStatus];


export const PageStatus = {
  draft: 'draft',
  published: 'published',
} as const;

export interface PageSeo {
  title: string;
  description: string;
  /** @nullable */
  ogImageUrl?: string | null;
  /** @nullable */
  canonicalUrl?: string | null;
}

export interface Page {
  id: string;
  title: string;
  slug: string;
  status: PageStatus;
  sortOrder: number;
  blocks: Block[];
  seo: PageSeo;
  updatedAt?: string;
}

export type PageInputStatus = typeof PageInputStatus[keyof typeof PageInputStatus];


export const PageInputStatus = {
  draft: 'draft',
  published: 'published',
} as const;

export interface PageInput {
  /** @minLength 1 */
  title: string;
  /**
     * @minLength 1
     * @pattern ^[a-z0-9]+(?:-[a-z0-9]+)*$
     */
  slug: string;
  status?: PageInputStatus;
  sortOrder?: number;
  blocks?: Block[];
  seo?: PageSeo;
}

export interface PageUpdate {
  /** @minLength 1 */
  title?: string;
  /**
     * @minLength 1
     * @pattern ^[a-z0-9]+(?:-[a-z0-9]+)*$
     */
  slug?: string;
  sortOrder?: number;
  blocks?: Block[];
  seo?: PageSeo;
}

export type PublishPageInputStatus = typeof PublishPageInputStatus[keyof typeof PublishPageInputStatus];


export const PublishPageInputStatus = {
  draft: 'draft',
  published: 'published',
} as const;

export interface PublishPageInput {
  status: PublishPageInputStatus;
}

export type MediaResourceType = typeof MediaResourceType[keyof typeof MediaResourceType];


export const MediaResourceType = {
  image: 'image',
  video: 'video',
  raw: 'raw',
} as const;

export interface Media {
  id: string;
  name: string;
  publicId: string;
  url: string;
  resourceType: MediaResourceType;
  /** @nullable */
  width?: number | null;
  /** @nullable */
  height?: number | null;
  createdAt: string;
}

export type MediaInputResourceType = typeof MediaInputResourceType[keyof typeof MediaInputResourceType];


export const MediaInputResourceType = {
  image: 'image',
  video: 'video',
  raw: 'raw',
} as const;

export interface MediaInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  publicId: string;
  url: string;
  resourceType: MediaInputResourceType;
  width?: number;
  height?: number;
}

export interface CloudinarySignature {
  timestamp: number;
  signature: string;
  apiKey: string;
  cloudName: string;
  folder: string;
}

export type ContactSubmissionStatus = typeof ContactSubmissionStatus[keyof typeof ContactSubmissionStatus];


export const ContactSubmissionStatus = {
  new: 'new',
  read: 'read',
  archived: 'archived',
} as const;

export interface ContactSubmission {
  id: string;
  name: string;
  email: string;
  message: string;
  status: ContactSubmissionStatus;
  createdAt: string;
}

export interface ContactSubmissionInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  message: string;
}

export interface ActivityItem {
  id: string;
  label: string;
  detail: string;
  timestamp: string;
}

export interface DashboardSummary {
  pages: number;
  publishedPages: number;
  media: number;
  submissions: number;
  recentActivity: ActivityItem[];
}

/**
 * Resource not found
 */
export type NotFoundResponse = Error;

export type DownloadSiteBackup200 = { [key: string]: unknown };